Add initial support for v_2_4
